chore: moved types to out main types file
